requirements.txt
python-dotenv==0.7.1
psycopg2==2.7.3
scipy==1.0.0
numpy==1.14.0
scikit-learn==0.19.1
